Cover The Tiles. The Dice Game
This is the Kutaisi International University's assignment. Java Courses: 
 you are going to implement a variant of the two-person dice game.
There are ten numbered tiles (from 1 to 10) which in the beginning, are all open.

Playing
Two player compete each other in every round. The respective player roles two dices and computes the sum of the obtained numbers. The player roles two dices. Player always rolls two dice and covers two tiles matching the sum of the dice roll. For example, ifthe dice show a 2 and a 3, player may cover one of 2 and 3 or 1 and 4. Selected ones are turned over. If no such option is available, the player's account is credited with the total of all open boxes. It is not necessary to automatically examine each possibility for the presence of at least one valid combination.

Game Over
There are a total of 10 rounds played, and each round consists of one option made by each player. Whoever has the smallest account balance wins. The player who successfully closes the final remaining boxes in a round wins instantly, regardless of the account values earned.